Move `Resolver.create` and `HasteMap.create` to `Runtime.create(Resolver|HasteMap)`.
This way `jest-haste-map` and `jest-resolve` don’t require knowledge of Jest’s configs and can be used more clearly outside of Jest.
